+++ examples/bind_all.rs
@@ -18,5 +18,5 @@ fn main() {
     });
 
     // Call this to start listening for bound inputs.
-    inputbot::handle_input_events();
+    inputbot::handle_input_events(false);
 }